Add extra Grafana dashboard capabilities (#50)
This makes it possible to define extra Grafana dashboards. We need this functionality for our customers as well as for the different circles.
Once you define one extra dashboard, it will also include all the other dashboards in the output. This is because of a limitation of the upstream Grafana chart where once you define an extra dashboard, you have to define all dashboards.