Add ability to customize alertmanager routes and receivers (#53)
We need to be able to customize alertmanager so that customers can add custom routes and receivers for their usage. This adds an extra terraform variable where you can add custom routes and receivers.
I added this a string instead of a map variable because the yaml of a route and receiver can contain a lot of custom parameters and nesting.