Update fluentd to version 1.2.1 and update its config to the new syntax (#65)
* Update fluentd to version 1.2.1 and update its config to the new syntax

* Remove #{Socket.gethostname} from the fluentd tags

Inside a pod, the hostname is the name of the pod. I haven't seen a
clear way to get the actual k8s node hostname from inside the pod using
env variables.

* Retrieve kubelet logs from systemd

* Remove duplicated fluentd log sources

These services are also logging as k8s services in the containers log,
which is also being picked up by fluentd, with much more context
information and metadata. So the logs where being duplicated.

* Get docker logs from journald

* 'filters' parameter is deprecated: filters has been renamed as matches